requirements.txt
firebase-admin
requests
Pillow
reportlab
pyserial
